requirements.txt
streamlit
yfinance
matplotlib
pandas
numpy
python-dotenv
langchain
langchain_groq
newsapi-python
textBlob 